database/requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
lxml>=4.9
beautifulsoup4>=4.12
pandas>=1.5
openpyxl>=3.1